Change png to svg export
Since the exported png image had a really bad resolution the format has
been changed to svg (scalable vector graphics). This allows the user to
zoom in and out without loosing any quality of the image, which is
really nice if you want to have a closer look to some month separately.