Adding GUI
After the image is generated a gui will appear to let the user see the
result. Although it is already working further features like multiple
language support or choosing the save directory are intended to be
added. In order to keep track of the project as a whole it is planned to
change the structure to something like MVC.